refactor(questionCard): move question components to question folder
