+++ src/views/main/Home.vue
@@ -410,7 +410,7 @@ import UserCard from '@/components/UserCard.vue';
 import ExploreSitesGrid from '@/components/ExploreSitesGrid.vue';
 import CommentCard from '@/components/CommentCard.vue';
 import ArticleColumnCard from '@/components/ArticleColumnCard.vue';
-import QuestionLink from '@/components/QuestionLink.vue';
+import QuestionLink from '@/components/question/QuestionLink.vue';
 import NewContentActionBar from '@/components/NewContentActionBar.vue';
 import Answer from '@/components/Answer.vue';
 import ExploreCard from '@/components/ExploreCard.vue';
@@ -419,13 +419,12 @@ import ExploreIcon from '@/components/icons/ExploreIcon.vue';
 import CloseIcon from '@/components/icons/CloseIcon.vue';
 import SharingIcon from '@/components/icons/SharingIcon.vue';
 import FeedIcon from '@/components/icons/FeedIcon.vue';
-import QuestionPreview from '@/components/QuestionPreview.vue';
+import QuestionPreview from '@/components/question/QuestionPreview.vue';
 import ArticlePreview from '@/components/ArticlePreview.vue';
 import UserGrid from '@/components/UserGrid.vue';
 import { readNarrowUI, readUserProfile } from '@/store/main/getters';
 import { setAppLocale } from '@/utils';
-import { dispatchCaptureApiError } from '@/store/main/actions';
-import { dispatchUpdateUserProfileQuiet } from '@/store/main/actions';
+import { dispatchCaptureApiError, dispatchUpdateUserProfileQuiet } from '@/store/main/actions';
 import { apiSubmission } from '@/api/submission';
 import CreateQuestionForm from '@/components/CreateQuestionForm.vue';
 import UIStyleControllers from '@/components/UIStyleControllers.vue';
